Resolved empty URL redirects, file_size handling and file details redirection
