+++ src/utils/fileSize.js
@@ -1,15 +1,15 @@
 const getFileSize = (file_size) =>{
-    if(file_size>=1024*1024){
-        return  (file_size/1024).toFixed(2)+" GB";
+    if(file_size>=1024*1024*1024){
+        return  Math.round(file_size/(1024*1024*1024))+" GB";
     }
-    else if(file_size>=1024){
-        return  (file_size/1024).toFixed(2)+" MB";
+    else if(file_size>=1024*1024){
+        return  Math.round(file_size/(1024*1024))+" MB";
     }
-    else if(file_size>1){
-        return (file_size).toFixed(2)+" KB";
+    else if(file_size>1024){
+        return Math.round(file_size/1024)+" KB";
     }
     else{
-        return (file_size).toFixed(3)+" KB";
+        return Math.ceil(file_size/1024)+" KB";
     }
 }
 export default getFileSize;
